[avr] Really truly working I2C chat code and STM32 USB interface!
It's been a journey! I've been on a roller coaster between elation and
despair for the past few days. First, excited about my simple tasking
code. Then, excited that I seemed to have finally written working I2C
firmware for the AVR.

But then when I tested it, via code running on an STM32 (which I was
communicating with via ST-LINK, which is notorious for hanging if you
look at it wrong, and requiring a power cycle), it CRASHED MY COMPUTER!
I got a Windows BSOD (blue screen of death). This happened twice.

That sent me reeling into despair. I thought "This can never work. There
are too many variables, too many moving parts: Windows, WSL, usbipd-win,
ST-LINK..."

On the advice of a friend I fired up WinDbg and pointed it at my Windows
crash dumps. It found the culprit: VBoxUSB, which is used in usbipd-win!
It turns out that zero-length USB packets can cause BSOD crashes - and
when I looked carefully at my code (to talk to the STM32 via ST-LINK) I
realized that I was reading *a lot* of zero-length packets!

The usbipd-win author issued a release that includes a fix to this
problem (to get it, upgrade to version 3.1.0 or higher).

I upgraded usbipd-win *and* fixed my code to never send or request
zero-length packets.

Then I decided to revisit the I2C code running on the STM32. It was
awkward and ugly in several respects, and I didn't have confidence in
it. I re-read the STM32F072 programming manual I2C section (several more
times) and re-wrote the code.

Everything went sideways again, and I freaked out! But with careful
and plodding debugging, I realized exactly where the problems were,
fixed them one by one, and now I have a very simple piece of code that
seems to work flawlessly.

The AVR will *always* NACK while executing a flash command (which can
take several milliseconds). Now the code that talks to the STM32 is
careful: if it gets a NACK, it waits 5ms and retries. It is written to
try up to 4 times, but it seems to never require more than one retry.

With all this in place, AVR devices can be reliably re-flashed over I2C!