[avr] Simpler (mostly) and more flexible multi-tasker
The inital one required that all tasks have the same stack size and the
same register save/restore conventions. This one has no such
limitations! And the user area is the same size. I re-used the byte of
"padding" between the status byte and the saved sp and turned it into a
byte-sized pointer to the next task's user area!

This way this multitasker is much like any other Forth multi-tasker:
the tasks are set up in a loop, and the tasker goes around and around
until it finds a task that is awake (interrupts can wake sleeping tasks,
so it's not a deadlock), and then restores that task's registers and
returns to executing it.