[avr] Removed signature region; rename app region to "flash"
When trying to connect the AVR memory code to the common flash code, I
realized that I had overly complicated things by adding a separate - and
actually useless - signature region.

I removed it and changed how the "application" code is laid out at the
beginning of the flash:

- vectors
- signature
- application

This actually simplifies everything.

Also, the "app" region is now called "flash", to be compatible with
everything else, including the common flash code.