[common] Flash code slightly more flexible
It's possible to call "verify-noisily", which *both* shows different
addresses *and* returns a difference value.

Also, the flash code indents the "erasing" and "programming" messages
slightly, so that if they are printed *after* a "global" message such as
"Updating application", they will be nicely indented underneath and
easier to read.