[avr] Added watchdog to serial chat example
I also moved some code around but I'm going to have to work on making
the chat code more modular.

I need to figure out how to separate out the list of task user areas and
the tasks woken by the heart/timer. Also, with the watchdog running,
there needs to be a way to stop it when reflashing the app code,
otherwise it will reset the chip in the middle of it.

But getting closer to a good "template" for a usable system.