[core] Improved termios support; revived serial-terminal
The termios code was a bit clunky. In order to make a change to a tty's
settings you had to execute *three* Forth words: get-termios,
set-termios-<something> (to make the change to the settings in memory),
then set-termios to actually change the termios in the tty device.

I decided to fold these three actions into a defining word, which also
captures an fd, so it's easy to create termios-changing words for either
the user's tty or for the serial port connected to a target device.

The reason for doing any of this is that I wanted to revive the serial
terminal code, which has been broken for a long time. I simplified it -
commenting out some weird special-case for the ARM AEB-1 eval board! -
and tested it against a "dummy" terminal (test/terminal-peer.mu4).