[pic18] WIP - updating to include newer Q series parts
The Q series - most or all of them - have a 6-bit BSR (bank select
register) and very different memory maps from the previous parts I was
dealing with (the 1xK50).

I've update the Instruction_formats, the assembler and disassembler, and
the memory management and meta-compiler.

But all of these are works in progress!

- they are untested
- the memory code is going to change to use the new "address-space"
  version of the common memory code

Once the "address space" memory code works for the 8051 and AVR I'm
going to return to the PIC18 and get it working there.

I want to get these changes committed so they don't get lost in the
shuffle. ;-)