[avr,8051] Converted AVR and 8051 targets to "aspace" memory model
Most of the targets that I have worked with in the past - ARM, HCS08,
RISC-V, MSP430 - have been von Neumann machines with a single address
space: ram, flash, and all i/o live in the same 16 or 32 bit address
space.

AVR, 8051, and PIC targets are different. They are Harvard machines,
with separate address spaces. The AVR has three address spaces: flash,
ram and i/o, and EEPROM. The 8051 has three as well: flash, ram, xram.
The PIC devices vary, but at the very least they have two spaces: flash,
and ram and i/o.

This difference caused a lot of trouble when trying to represent the
state of the target's memory in host-resident memory images. My model
previously was to have multiple memory *images* - generally two: flash
and ram - and in each image to have multiple *regions* - these are like
sections in a Unix binary, witth a starting address and length.

The aspace - address space - model adds an abstraction above these two.
Every image belongs to a single aspace, but there can be multiple images
in an aspace. Every region belongs to a single image, but there can be
multiple regions in an image.

Thus, the current *region* also defines a current image and current
aspace.

Eventually I want to move *all* the targets to this model; the von
Neumann ones simply have a single aspace!