Vim, again: I'm essentially reverting to commit 5aa61892
I'm bouncing back and forth between that commit - which treats all
printable characters except parens, hyphens, and underscores, as keyword
characters - and commit babf0ad7, which treats all printable characters,
with no exceptions, as keyword characters.

This matters when using "w" and friends to move around in Vim. If all
printable characters are keyword characters, then "w" will jump from one
Forth word to the next.

If we exclude parens, hyphens, and underscores, then parts of compound names are
easier to change, and parens are commonly used to surround the names of
"runtime" words.

It's a mixed bag. Not sure what to make the canonical version!