[core] Exiting with "bye" no longer messes up the console
Changes to the tty code in October 2023 could leave the console in raw
mode, so that when the user, sitting at their shell prompt, typed
anything, it wasn't echoed. Blindly typing "stty sane" (if you knew to
do that!) would return the console to a usable state, but now muforth is
more careful about when the console is in raw mode versus cooked mode.