[avr] Fixed a typo that prevented default device from working
I changed the default from atmega168 (the original Arduino chip, but now
hard to find), to atmega328 (the current chip on the Arduino Uno and
probably other clones, easy to find).

When I made the change I clobbered a closing quote ("), causing the file
to fail to load. This fixes that.