[many] Got rid of "kick" in various target "interact" files
This all started because the word "run" was taken by the ARM ST-LINK
support code. So I renamed "run" to "kick" in the ARM/v6-m/interact
code, and it infected several other targets.

I put a prefix of "dbg." on a bunch of words in the ST-LINK code,
renamed ARM interact's "kick" back to "run", and did the same for
several other targets that were affected.

AVR isn't in the list, however: there are a bunch of AVR changes still
coming that involve the interact code, so that will have to wait.