[avr] Big changes to serial chat and interact code!
I starting by trying to save space - so the chat code would fit into 512
bytes - and ended up realizing that chat entry and running code on the
target could be made much simpler and more flexible.

This is, in a sense, a work-in-progress: the tasking code can also be
changed and simplified, and I need to figure out how the "trampoline"
code will work, both for executing "native" and "Forth" code (even
though I don't really think there is a difference: Forth on the AVR will
have two stacks, but will *not* be threaded).