[avr] Hopefully fixing the serial (uart) chat code
For 8M RC oscillator, run UART at 38400 instead of 57600, as the
effective baud rate is much closer to the standard one.

Push and pop the flash read count register around calls to yield; all of
the registers are caller-save as far as the multi-tasking code is
concerned. The count was probably getting clobbered.

And a bit of weird trivia: 9600 hex equals 38400 decimal!