[avr] Added "register style" remote execution
By adding to the application code a "trampoline" that pops several
registers from the data stack (pointed to by Y), executes a piece of
code, and then pushes the same registers back onto the data stack, we
have an easy way to try out pieces of code.

Later we can add a more "Forth-oriented" trampoline, but it will be very
similar to this one.

In addition to the trampoline, there is some example code to try out.
Look in target/AVR/register-examples.mu4.

To try it out, load everything like this:

  ./muforth -f target/AVR/load-chat-serial.mu4
  prog
  chat

Put values into registers like this:

  abcd w!
  0200 x!

Except for SREG, SP, and PC, the registers shown in the stack display
can all be accessed this way.

Then execute your code:

  ! rx

The resulting registers and SREG will be displayed, and results can be
fetched usign w@ x@ etc.