[avr] Added a "dummy" chat target - useful for testing
I wanted a way to see how the host would interact with the new "register
trampoline".

The chat dummy - similar dummy targets exist for the S08 and MSP430
targets - "pretends" to be target, and can do everything except execute
code. Thus, programming the flash, copying values into the RAM, all
these appear to work just as they would on a real target.

To load and try it, do something like this:

  ./muforth -d dummy -f target/AVR/load-serial-chat.mu4