[avr] WIP - remote exec - one last gasp!
I think I figured out what's going on. I was mostly likely executing the
code properly, but I couldn't get anything back from it because it was
storing its context into register Y, which promptly gets clobbered by
the multitasker after the code has completed executing, and chat is
waiting for the next command.

What I did on this commit is a bit crazy - I'm using Z for *both* the
address of the code to execute (which the host sends in return address
(ra) form: the flash address, shifted right one bit, with the bytes
swapped) AND the stack (or context) pointer! I'm just not using them at
the same time.

Here's how it works: the host uses the set-address chat command to set Z
to the ra. It then executes the run chat command, which pushes Z onto
the stack and then falls through to set-address; the host then sends the
stack pointer, which is put into Z, and the ret instruction in
set-address actually calls the code!