[avr] Remote execution (via register trampoline) working!
We now use the Z register for everything!

To try it out, first re-flash the chat code, then do this:

  ./muforth -f target/AVR/load-chat-serial.mu4
  chat
  7fc w! @ rx

and you should see "cafe" in W, which is the value of the stack sentinel
(which is always at 07fc in ram).

Use w! x! z! h1! to store values in registers; w@ etc to fetch them.
Find example code to execute in target/AVR/register-examples.mu4.