[arm] Move STM32 flash programming routines to below stacks
They were at the RAM origin (@ram) but this can be in the way of code or
data structures that we might want to put there.

Moving the flash routines higher - so that they are just below the
return and data stacks - and copying them separately from the ram region
(which could be empty - the flash routines don't affect the ram region
pointers) - makes it easier to use the ram as we please.