[arm,avr] STM32 AVR programming code runs from flash or ram!
I decided initially that it would be "easier" to always run the
programming code from ram. This meant that each time you wanted to
program an AVR you had to load the STM32 code onto the board, start it
running, and then load the AVR code that you were interested in. Yes,
you could leave the STM32 in the background, running its ram-based code,
but having it be two steps to begin with is maybe not ideal.

Now you have a choice!

To load the code into ram and start it up, do this:

  ./muforth -d in-ram -f target/AVR/load-stm32-discovery.mu4
  startup run
  ^C

To load the code into flash, do this:

  ./muforth -f target/AVR/load-stm32-discovery.mu4

If there is *no* message saying that the flash images don't match, then
the correct code is already in the flash. You can ^C to exit from
muforth, but since the ST-LINK has halted the CPU, you need to reset the
STM32 Discovery board (press the black switch).

If you *do* see the message that the flash images don't match, then you
should do this (while still connected to the ST-LINK):

  flash-image
  verify
  ^C

and reset the board, as above, by pressing the black switch.

That's it!