[avr,8051] Programming via SPI: use hardware SPI instead of bitbanging
I left the bitbang code there as an option, but the new default is to
use the STM32's hardware SPI peripheral.

I was frustrated by having to do ugly delays to get the "slow clock"
version of the code - necessary for programming the fuses on "factory
fresh" AVR chips, which are only running at 1 MHz - to work.

By using the SPI peripheral I can easily change the clock divisor to get
the slow or fast clocks needed.

For the 8051 (really the AT89LP) version of the code, use an even faster
clock (3M). (It can be 5M max, but we can't get that with the way the
STM32 divisors work.)