[host] Made "endian-aware" fetch and store words more useful
It struck me that an endian-aware fetch or store word like "lew@" -
which, given an address, fetches a 32-bit little endian word value from
memory - is not quite the right primitive. A better choice is its
*incrementing* sibling "lew@+", which leaves an incremented address on
the top of the stack.

The non-incrementing version is then simply defined as

  : lew@   lew@+ drop ;

Why is this useful? It's often necessary to fetch a series of values out
of a *buffer* that one has received over a wire from a target. The sizes
of the value can be mixed by using different fetch operators.

Let's say we want to fetch two little-endian words, followed by two LE
halfwords (16 bit values), from our buffer. Simply do this:

  buffer  lew@+ lew@+  leh@+ leh@

Similarly with storing, and big-endian values. If we have the following
values on the stack - d c b a - we can store them as two big-endian
words followed by two big-endian halfwords, thusly (the comment shows
which stack value gets stored by each operator):

        (  a     b      c     d)
  buffer  bew!+ bew!+  beh!+ beh!

In both the fetching and storing case, the *last* operator is a
non-incrementig version, which consumes the address.

I'm not sure why I never thought of doing this before! It's quite
useful, and rather elegant.