[pic18] Added comments to old -K ICSP programming code
This was used for programming the 13K50 and 14K50, which I was
interested in for their USB peripheral.

They use a complicated technique that involves clocking instructions
into the device and executing them, one by one! The whole process was
orchestrated by code running on an S08 device - in the same way that I'm
now using STM32 ARM devices (Discovery boards).

This code is not useful as is, and needs some work. Ideally it should
be made to work using an STM32.