[many] Vastly improved and generalized flashing and verifying code!
All the targets except for MSP430 and S08 are now using this.

It has helpful diagnostic messages, and is very customizable via
defer'ed words.

I want to have common code for all of the following:

- memory images and regions
- interaction with the target
- meta-compiler
- flashing and verifying

So far I'm pretty close on the first and last. Number three is the
hardest!