[pic18] Working Flash and Config/EEPROM programming via chat!
This means that flash-image and verify will now work when chatting via
UART with the target.

Note that the boot region is protected - via hardware - from erasure or
re-programming. This is where the chat code lives, so we wouldn't want
to clobber it!

Even so, I've added boot to the list of regions that get verified. So if
there *is* anything in boot, it will be compared with the target's boot
region. This might end up being helpful.