[pic18] Assembler and disasm simplified and improved (hopefully)
- disasm prints equates, labels, and target addresses by name

- the "effective address" keys - "c" (call), "g" (goto), and "r"
  (return) now work in the disasm

- asm and disasm: I have dropped the x y z names for FSR0, 1, and 2, and
  generalized the load, add, and subtract instructions (lda, adda, suba)

- both: renamed the "tbl" instructions to "prog"

Initially I thought that using the names x, y, and z instead of FSR0,
FSR1, FSR2 would be better, but then, as I was writing the kernel words
for the PIC18, I wanted to use X for a scratch stack slot, so I needed
something else.

I went back to using the generic names (FSR) in the asm and disasm, but
added the capability (via a defer'd word in the disasm) to print them
however the user wants.