[pic18] WIP: Untested, unfinished support for 18F1xK50 devices
Years ago I was able to get a small chat program onto the 13k50/14k50 -
enough to poke around and dump memory but nothing more.

I programmed the chips via a quirky S08-based system (that I don't
expect anyone to duplicate), and so for broader consumption I need to
port this code over to the STM32 - a future project.

This commit makes the recent chat code (which works great on the
Q-series PIC18 devices) at least *compile* for the K50, but I have not
tried programming it into the flash and trying it out.