[arm] Flash programming supported on Nucleo-C031C6 board
The STM32C series chips are unlike the F0 and F3 parts in the way that
they deal with erasing the flash. Instead of using an address - the
address of the first byte of the page of flash to erase - they use an
*index*.

The C031 has 32 KiB of flash, divided into 2 KiB pages (or sectors).
Thus the index is from 0 to 15.

The flash i/o registers are laid out like the F407, which also uses
indices to erase the flash, but it has non-uniform sector sizes, and
they are quite large: I think the smallest one is 16 KiB.

So the C031 falls in between. It uses small uniform pages, like the F0
and F3 parts, but erases (and programs) them more like the F407.

Go figure. Something always has to change!