[core] [WIP] 32-bit heap: addresses on stacks are heap-relative
My end goal is to have a heap that on both 32-bit and 64-bit
architectures contains only 32-bit cells.

In order to do this, any cells that contain addresses - such as link
fields and colon word bodies - must contain *heap-relative* addresses.

A first step toward this goal is to modify the system so that any
address that is on the data or return stack is heap-relative rather than
host-absolute.

On 64-bit machines, this means that the stacks have to be 64 bits wide,
because a heap-relative that points a random mmap buffer, or indeed
points into the stacks themselves (which are staticallly allocated),
might be a very large offset, easily exceeding 32 bits.

As of this commit, the core muforth compiles and runs. It is a slightly
odd hybrid system in which variables contain heap-relative addresses of
code, but the bodies of words contain host-absolute address. But it
works perfectly, and the changes were relatively modest.

I have not tried to load or use any of the target compilers; that is the
next step.

After that, it should be fairly easy to change the muforth core to use
32-bit heap-relative addresses.

I have struggled for a long time to find an elegant approach to an
implementation that feels and works the same on 32- and 64-bit machines.
I think this will be it!