[core] [WIP] 32-bit heap: It works! (except for exceptions)
This commit is "part two" of the change to a pure, relatively-addressed,
32 bit heap. Part one was making sure the system worked when passing all
addresses on the stack as heap-relative ("heapified") addresses. (Note
that this requires a 64-bit wide stack on 64-bit machines.)

What *doesn't* work is the exception handling.

The return stack saving and restoring of values works fine, but
processing an exception and unwinding the stack does not. I think the
unwinding part is probably fine; what is broken is catch and throw.

So that's next.

Then, I may remove  compile,  because it is identical to the regular  ,
.

Lastly, I need to make sure that the target compilers all still work.
Probably a few minor changes will be necessary.