[core] [WIP] 32-bit heap: Signedness challenges!
Since the heap contains 32-bit cells, and the stack is 64-bit cells, we
have a question: when pushing a 32-bit value onto the stack, do we
sign-extend or zero-extend? This is equivalent to asking: are values in
the heap (or literals) signed or unsigned?

The problem is, they need to be both. Mostly unsigned, but sometimes
signed. This is annoying, and it needs to be addressed.

This commit addresses the problems by creating two kinds of literals and
constants (signed and unsigned). The word  @  by itself does an unsigned
"load" but can be followed by  s>64  to sign-extend to 64-bits.

This problem is "fixed", but the existence of the problem and the fact
that someone might try to store a stack value that needs more than 32
bits into a 32 bit heap cell made me re-think this whole series of
changes!

While everything seems to work - although there was a quirk that I found
when verifying addresses on a PIC18, using chat, it printing addresses
like 1_0000_0040 - I may shelve this and end this branch here.

I'm tempted to try a 64-bit *relative* heap version - so heap cells and
stack cells are the same size again - and if having a relative heap
doesn't seem very useful, go back to the beginning of muforth being
64-bit with a simple uniform codebase, perhaps this time with no
"special treatment" for 32 bit machines, which would run the code a bit
slower.