[core] Removed addr@, addr!, addr-preserve
These were used for only one purpose: to read and write two significant
C pointers in the text interpreter:

  static char *first, *end;

Since C pointers are 32 bits on 32 bit machines, and 64 bits on 64 bits
machines, we had to treat them specially; hence, the "addr<something>"
words.

By creating a "value-based" (getter/setter) interface to first and end,
we can remove the addr words.

  source@  pushes the values of first and end;
  source!  takes new first and end values from the stack and stores them
           into the C variables.

In order to save and restore the source pointers - ie, when loading a
file - we need a way to execute a piece of code, on exit from a word,
that takes *two* stack values.

We already had  on-exit  which takes *one* value, so I just created an
on-exit2 and used it to save and restore the source.

While it's slightly weird to have code that is only used in one place,
it's actually totally generic code, and might be useful elsewhere.