[various] All the targets now use the same memory image code!
I had to touch a lot of files, including the interact code for many
targets.

I also added an easy way to set up the memory image code for 16- and
32-bit von Neumann targets. Look at

  mu/target/common/memory-von-neumann-*

I left the ifdef'd "legacy" code in place. Going in and deleting it is
probably a useful task. I'll put that in its own commit so it's easy to
find the vanished code.