[core] Lua script to improve multi-line comments
I have been trying to understand Vim's handling of "three-part" comments
- with a start, middle, and end - and finally figured out how to make
them work.

I had been using a Vim three-part comment definition in which the middle
part consisted only of whitespace, but this doesn't work very well.

To fix this problem, I tried a version whose middle part starts with
a vertical bar or pipe character ("|"), which aligns with the left
parenthesis that began the comment, and found that this works really
well and creates a block comment with a vertical line running down the
left side, which helps to distinguish comments from text.

With the proper definition, it's also really easy for Vim to
successfully reformat comments, regardless of their indentation.

Overall, it's really great.

The problem is, there is a whole tree full of code that is in the old
style, and converting it by hand would be painful.

Hence this script. One nice thing about it is that it's idempotent.
After running it on the tree, running it again does nothing, so it can
be easily used to keep a tree of code in good shape, without worrying
that previously converted files will get screwed up.

It's easy to do. cd to the top-level muforth directory and run

   find . -name "*.mu4" -exec scripts/do_lua.sh {} \;

git diff will show you what has changed.

I haven't yet committed the new Vim "ftplugin" file for mu4 files yet.
That is coming soon!