[core] Updated comment rewriter Lua script
There are two ways to do block or multi-line comments in muforth:

(1) Surround a block of text with "( " and ")". This was the original
change: making all lines except for the first start with "|".

(2) Using the "comment" word. It consumes the following token, and then
goes into a loop, throwing away tokens until it finds the token again.

I added support in the rewriter script for (2), and in so doing had to
do some iteration. Several of my uses of "comment" involved using
a token consisting of a horizontal line of chars that matched - and
would align with - the closing use of the token.

Because I am adding "| " to the beginning of every line between the two
appearances of the framing token, the alignment can suffer. So I had to
be careful about that.

At this point, I think the script works well for both types (1) and (2).