Updated comment rewriter script (again)
It now has subcommands: block, dashdash, brackets, parens.

I'm about to run the block subcmd on the whole tree and check in the
changes that result. Then I will run the others in order. The script
might need updating along the way (getting everything right is
iterative, and their are some weird "exception" cases that have to be
found and dealt with).

By the end we will have a neat new tree with spiffier, easier-to-read
comments.