All .mu4 files processed with comment rewriter "brackets" subcommand
Because I was using ( ) to delimit block comments, any ")" in the
comment text would end the comment. Thus I used [ ] to delimit
"parenthetical" remarks.

This pass rewrites those [ ] as ( ).

The problem is, there are a bunch of false negative matches, so the
matching code has several exceptions in it (and is very ugly and was an
iterative nightmare to arrive at). I could never get it to be perfect,
and had to undo (using git restore -p) several of the changes that the
script made before committing.

I don't recommend running this on a tree without a lot of care. While it
only changes comments - lines that start with optional whitespace and
then the "|" character - it might really mess them up and/or change
their semantics! Here be dragons!

After this commit, I am done messing about with comments.

Time to get back to doing some Real Work (tm)!!

More seriously, I think code readability is deeply important, and
I don't mind having taken the time and care to work on this.