[arm] Fixed STM32 flash programming code; improved target exec
When I added support for the STM32C031 (commit b0ca148c), I broke the
flash programming code for F0 and F3 devices!

For some reason I decided it was a good idea to copy over some code - which
cleared some error bits - from the C031, but there is a big
difference between the F0/F3 and the F4/C0 devices: the flash registers
on the F0 and F3 devices can *only* be accessed as 32-bit values (using
ldr and str), while the F4 and C0 devices can be accessed as byte,
halfword (16 bits), and word (32 bits).

The code I copied used ldrb and strb, and this caused the devices to
basically go "splat" and act very unpredictably. They crashed in a very
mystifying way!

I found the problem, and decided, while fixing it, to do two additional
things:

  - refactor and simplify the code (now it uses only the hardware stack,
    so calling it is easier);

  - add a *timeout* when executing code on the target, instead of
    hanging (which also causes ST-LINK to flip out, and often requires
    power-cycling the board).

I refactored and simplified the F4/C0 code as well, but, sadly, I don't
have hardware to test it. (At the moment, anyway.)

In addition, since I changed the API for "get-status" - it now returns
a "retry count" that represents how many times the target was polled for
completion while executing code - so I added support for it to all the
files in target/ARM/debug/. The only other interesting one is probably
the CMSIS-DAP driver.

I think I'm probably going to merge the "uniform page" and "sectored"
code into one file. Much of the code is shared, and duplicating the
shared parts is annoying. I had to do a lot of copy and paste for this
change.