Vim-related: Improve installation of Vim files into $HOME
Previously the "install-vim" makefile target would tar up the .vim/
directory in the muforth source tree and untar it into the user's home
directory.

This works fine until the muforth settings change.

It seems like it might be easier to make symlinks from the user's .vim/
to muforth's .vim/, so the user gets updated when muforth updates.
(Unless they don't want the muforth changes, in which case they have to
break the symlinks and copy and modify the files manually.)

To make this work easily I had to change a shell variable. I used to use
"MU_DIR" to point to /some/path/muforth/mu/. Since the .vim/ directory
is above mu/, now I have MUFORTH_ROOT, which points to
/some/path/muforth/. I can use this for the (modified) install-vim
target, and MU_DIR is simply MUFORTH_ROOT/mu/.