[avr] App entry point can be used to start and stop
Previously the chat/bootloader would call the app entry point (the RESET
vector) to have the app create and init its tasks. There was no way to
have tell the app to stop and shut down (presumably, before re-flashing
a new version of the app).

Now the app checks the h2 register. If non-zero, it creates tasks; if
zero, it shuts down.