[common] Flash verify improved (granular accuracy)
It sometimes got the wrong answer at the edges (generally the end) of
a region being verified because it was doing the comparison a host cell
size (8 bytes) at a time before "zooming in" on the differences at the
granularity specified.

Now there are three loops, one for each granularity (1, 2, or 4 bytes).

There is a slightly better way, and that's coming next. But I wanted to
snapshot this fix.