[arm] Pico: renamed XIP_SSI_* to SSI_*
The equates changed in the SVD file that we use to generated them, and
I updated the equates file (in commit b2072a39), but never touched the
code there. It turns out the XIP experiments I had done in my early days
of exploration with Pico would no longer load because of the name
changes. These are now fixed.