[avr] Finally completely fixed I2C chat "restart" implementation
I needed a way to clear the return stack and return to the chat command
loop when the slave device was re-addressed. Everything I have tried
until now has been flaky. This approach is very simple: I reset the
stack pointer based the task user area address (the stack is immediately
below this), and I use the "hook" support for creating a forward
reference to the place where we wait for a command byte from the host.

It works perfectly.