[avr] Simplified example register trampoline code
Instead of having local .W .X .Z variables that you have to store values
into before executing a piece of code, why not just leave the values on
the stack?

And *after* executing the code, why print them out? Why not leave them
on the stack?

This way, once the stack has three values - for w, x, and z registers
- you can simply do, eg

  negate rx

and the right thing will happen.