[avr] Finally, a sensible naming convention for stack pointers!
After much tearing of hair and gnashing of teeth, I have decided on the
following naming convention for stack pointers:

  DP  -  data stack pointer
  RP  -  returns stack pointer
  MP  -  machine stack pointer (used for call/ret, push/pop, etc)

Note that *none* of these is called "SP"!! That is by design. When you
see one of these you should stop and think: what is this? What does it
mean?

The reason that this matters is twofold:

  * Forth code, and code written in a Forth "style", uses *two* stacks:
    one for data items and one for (nesting of) return addresses;

  * Threaded Forths use the machine stack for data and a general-purpose
    pointer register for the return stack, while native code and
    non-threaded Forths are the reverse!

This gets confusing, especially when setting up the registers to begin
code execution.

Having distinct names for these three concepts is very useful!

I've converted the AVR code to use this (though mostly that meant
changing from CP (context pointer) to DP (data stack pointer).

Eventually I want all the targets to use this convention.