[s08] Big changes to S08 chat and interact!
I was inspired by some things I had learned recently while working on
the AVR code and wanted to try to apply them to the S08. It's also been
a *very* long time since I have touched the S08 code, so I thought it
would be fun.

I have always enjoyed programmed the S08. It's a very simple and limited
processor, but it has a certain 1970s flair that I really enjoy. More
than perhaps any other microntroller I have used, programming the S08
often feel like magic! It's hard to explain. But, even after a gap of
several years, I'm still feeling that "magic".

Anyway, I have been struggling with the "right way" to execute code on
a target device in a tethered environment. I think the S08 was first
processor I tried doing this with in muforth. I decided to use the SWI
instruction, and to base everything off a "chat frame" - containing the
register state and condition codes - that get pushed (by executing swi)
- every time chat is entered.

To execute a piece of code, useful register values are put into the chat
frame (there is a local copy on the host), and then the chat frame is
copied over to the target, and a reti (return from interrupt)
instruction is executed, which pops all the registers and then runs the
code pointed to by the PC.

This works great, but as I started executing Forth code more, rather
than assembly/machine code - this started in earnest with the MSP430
- I realized that the chat frame is not that helpful. In general you
want the flexibility of running an external "trampoline" that will
marshal arguments in and out of registers and set up the environment for
running Forth.

Basically what was needed in chat was much simpler than having
a snapshot of the registers that is passed back and forth. Instead, the
chat code passes just two values to the target - a pointer (generally
speaking, this is to a data stack) and a PC to jump to. The pointer is
put into a suitable register, and the PC is jumped to - with an address
on the return stack for re-entering chat. The PC points to
a *trampoline*, which moves values between the stack and registers as
needed, both before and after executing *another* piece of code, whose
address is on the stack (eg, a Forth cfa).

I have used this model extensively on the MSP430 and the ARM Cortex-M
and it works very well.

It turns out it works just as well on much "smaller" processors too,
like the AVR and S08.

Using this idea - which moves code *out* of chat and into the
trampolines - it's possible to have a chat/bootloader on the S08
- capable of reading and writing ram, reading and programming flash, and
  executing code - which is 139 bytes long (including the code to
initialize the clock and UART).

All the complexity of calling code is moved outside of chat and into
a trampoline, and on a von Neumann processor like the S08 this means the
trampoline can be RAM-resident, which is quite nice. It's also very
useful because a well-written trampoline essentially makes the target
feel like an extension of the host machine: the normal Forth stack is
used to pass everything - including register values - back and forth.

In the process of doing all this I decided to fix, once and for all,
a complicated and ugly piece of code that copies those stack values back
and forth. It existed in several forms (approximately one per target)
and none of them was very pretty or readable.

I have finally created a single reusable version, and the S08 target is
the first to use it! The others should follow, as I end up working on
them.

Sadly, NXP has decided to mostly bury the S08. The parts that I really
liked - in particular the JM32 and QE32 - are now more expensive than
much more capable ARM parts, and nothing is available in breadboardable
DIP form. The S08PT and S08PA are newer and cheaper and maybe worth
exploring. But it's a bummer that you can't get a chip, stick it in
breadboard, and start coding. The world has moved on.

Sorry about the novel. I hope you enjoyed reading it.